move codes to replace
